+++ src/app.rs
@@ -582,6 +582,7 @@ pub fn all_args_and_flags() -> Vec<RGArg> {
     flag_line_number(&mut args);
     flag_line_regexp(&mut args);
     flag_max_columns(&mut args);
+    flag_max_column_preview(&mut args);
     flag_max_count(&mut args);
     flag_max_depth(&mut args);
     flag_max_filesize(&mut args);
@@ -1440,6 +1441,30 @@ When this flag is omitted or is set to 0, then it has no effect.
     let arg = RGArg::flag("max-columns", "NUM").short("M")
         .help(SHORT).long_help(LONG)
         .number();
+    args.push(arg);
+}
+
+fn flag_max_column_preview(args: &mut Vec<RGArg>) {
+    const SHORT: &str = "Print a preview for lines exceeding the limit.";
+    const LONG: &str = long!("\
+When the '--max-columns' flag is used, ripgrep will by default completely
+replace any line that is too long with a message indicating that a matching
+line was removed. When this flag is combined with '--max-columns', a preview
+of the line (corresponding to the limit size) is shown instead, where the part
+of the line exceeding the limit is not shown.
+
+If the '--max-columns' flag is not set, then this has no effect.
+
+This flag can be disabled with '--no-max-column-preview'.
+");
+    let arg = RGArg::switch("max-column-preview")
+        .help(SHORT).long_help(LONG)
+        .overrides("no-max-column-preview");
+    args.push(arg);
+
+    let arg = RGArg::switch("no-max-column-preview")
+        .hidden()
+        .overrides("max-column-preview");
     args.push(arg);
 }